backend/requirements/base.txt
Django==5.1.2
djangorestframework==3.15.2
psycopg[binary]==3.2.12
python-dotenv==1.0.1
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.4.0
